[FLINK-33226][table] Forbid to drop current database

+++ flink-table/flink-table-api-java/src/main/java/org/apache/flink/table/catalog/CatalogManager.java
@@ -60,6 +60,7 @@ import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.Predicate;
@@ -1404,6 +1405,10 @@ public final class CatalogManager implements CatalogRegistry, AutoCloseable {
     public void dropDatabase(
             String catalogName, String databaseName, boolean ignoreIfNotExists, boolean cascade)
             throws DatabaseNotExistException, DatabaseNotEmptyException, CatalogException {
+        if (Objects.equals(currentCatalogName, catalogName)
+                && Objects.equals(currentDatabaseName, databaseName)) {
+            throw new ValidationException("Cannot drop a database which is currently in use.");
+        }
         Catalog catalog = getCatalogOrError(catalogName);
         catalog.dropDatabase(databaseName, ignoreIfNotExists, cascade);
         catalogModificationListeners.forEach(
